storage_types: extract RegionInfoProvider
