+++ components/storage_types/src/region_info.rs
@@ -2,6 +2,7 @@
 
 use kvproto::metapb::Region;
 use raft::StateRole;
+use std::error;
 
 #[derive(Clone, Debug)]
 pub struct RegionInfo {
@@ -16,3 +17,24 @@ impl RegionInfo {
 }
 
 pub type SeekRegionCallback = Box<dyn Fn(&mut dyn Iterator<Item = &RegionInfo>) + Send>;
+
+pub trait RegionInfoProvider: Send + Clone + 'static {
+    /// Find the first region `r` whose range contains or greater than `from_key` and the peer on
+    /// this TiKV satisfies `filter(peer)` returns true.
+    fn seek_region(&self, from: &[u8], filter: SeekRegionCallback) -> RipResult<()>;
+}
+
+quick_error! {
+    #[derive(Debug)]
+    pub enum RipError {
+        Other(err: Box<dyn error::Error + Send + Sync>) {
+            from()
+            cause(err.as_ref())
+            description(err.description())
+            display("unknown error {:?}", err)
+        }
+    }
+}
+
+pub type RipResult<T> = std::result::Result<T, RipError>;
+
